Transaction Complete..

+++ src/main/java/com/luuzun/ksca/service/BranchServiceImpl.java
@@ -8,13 +8,17 @@ import org.springframework.stereotype.Service;
 
 import com.luuzun.ksca.domain.Branch;
 import com.luuzun.ksca.persistence.BranchDAO;
+import com.luuzun.ksca.persistence.SccDAO;
 
 @Service
 public class BranchServiceImpl implements BranchService{
 
 	@Inject
 	private BranchDAO dao;
+	@Inject
+	private SccDAO sccDao;
 
+	
 	@Override
 	public List<Branch> listAll() throws Exception {
 		return dao.listAll();
@@ -36,8 +40,21 @@ public class BranchServiceImpl implements BranchService{
 	}
 
 	@Override
-	public void delete(String areaCode, String branchCode) throws Exception{
+	public String delete(String areaCode, String branchCode) throws Exception{
+		
+		//99 분회는 삭제 불가
+		if(branchCode.equals("99")){
+			return "ERROR:def";
+		}
+		
+		//분회에 소속된 경로당이 존재하면 삭제 불가
+		if(sccDao.readByBranchCode(areaCode, branchCode).size()!=0) {
+			return "ERROR:cascade";
+		}
+		
 		dao.delete(areaCode, branchCode);
+		
+		return branchCode;
 	}
 
 	@Override
